iio: imu: st_lsm6dsx: always flush fifo on probe/shutdown
Several platforms provide IMU with always on power.
The reset of the IMU does not reset the fifo and its interrupt.
Thus creating a racing condition when the irq is registered.
Make sure the fifo is flushed on probe and on shutdown.
Additionally, disable IMU sensors on shutdown to conserve power on such platforms.