power: supply: bq2419x: remove dependency to fuel gauge
Before a dependency was added to defer probe until fuel gauge is ready.

That was to avoid a hang caused in thermal core because it not initialized yet.
(Thermal core in fs and charger in subsys. Moving charger into fs sync would also fix that.)

Now thermal is initialized in core so there is no need for that work around.