phy: tegra: xusb: add usb3 no mapping WAR
Boards that need USB3 disabled and only have one port, do not work on
device mode on non USB2-only host ports.

That's because for plug detection to work on USB3 host ports, at least one USB2
port must be mapped to a USB3 port.

T210B01 does not have that issue, since it's missing port3 and initializes it
on boot and SC7 for power saving.

For T210 boards, a new dt u32 property was added to enable the workaround:
`usb3-no-mapping-war`